% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanic.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Titanic Data}
\format{
A data frame with 887 rows and 8 variable
}
\usage{
data(titanic)
}
\description{
Real data from the Titanic passenger liner disaster in 1912. Thanks to Kaggle and encyclopedia-titanic for the dataset.
}
\details{
\itemize{
\item{survived}{Indicator of whether the passenger survived}
\item{pclass}{Passenger-class}
\item{name}{Passenger name}
\item{sex}{Passenger sex}
\item{age}{Passenger age}
\item{siblings_spouses_aboard}{Number of siblings/spouses aboard}
\item{parents_children_aboard}{Number of parents/children aboard}
\item{fare}{Fare paid (pounds)}
}
}
\references{
https://www.encyclopedia-titanica.org/ \cr \cr
https://web.stanford.edu/class/archive/cs/cs109/cs109.1166/problem12.html
}
\keyword{datasets}
